# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactKind": {
        "enum": [
          "audio_original",
          "audio_enhanced",
          "audio_rendered",
          "midi_performance",
          "midi_corrected",
          "musicxml_score",
          "rendered_score",
          "stems",
          "analysis_report"
        ],
        "title": "ArtifactKind",
        "type": "string"
      },
...
      "Capability": {
        "properties": {
          "accepted_input_kinds": {
            "items": {
              "$ref": "#/components/schemas/ArtifactKind"
            },
            "title": "Accepted Input Kinds",
            "type": "array"
          },
          "failure_modes": {
            "items": {
              "type": "string"
            },
            "title": "Failure Modes",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "produces_output_kinds": {
            "items": {
              "$ref": "#/components/schemas/ArtifactKind"
            },
            "title": "Produces Output Kinds",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "version"
        ],
        "title": "Capability",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "cache_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Key"
          },
          "capability": {
            "$ref": "#/components/schemas/Capability"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_details": {
            "additionalProperties": true,
            "title": "Error Details",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_version_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Input Version Ids",
            "type": "array"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/JobLifecycle"
          },
          "output_version_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Output Version Ids",
            "type": "array"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "provenance": {
            "additionalProperties": true,
            "title": "Provenance",
            "type": "object"
          },
          "workflow_id": {
            "format": "uuid",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "capability"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobLifecycle": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "current": {
            "$ref": "#/components/schemas/JobStage",
            "default": "queued"
          },
          "lease_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Expires At"
          },
          "max_retries": {
            "default": 3,
            "title": "Max Retries",
            "type": "integer"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "progress": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "retry_count": {
            "default": 0,
            "title": "Retry Count",
            "type": "integer"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/ProcessingStatus"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          }
        },
        "title": "JobLifecycle",
        "type": "object"
      },
      "JobStage": {
        "enum": [
          "queued",
          "claimed",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStage",
        "type": "string"
      },
...
      "ProcessingStatus": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "progress": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "stage": {
            "$ref": "#/components/schemas/JobStage"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          }
        },
        "required": [
          "stage"
        ],
        "title": "ProcessingStatus",
        "type": "object"
      },
...
      "TextPassageFindWorkflowBody": {
        "description": "Queue one text query over one exact audio/performance Version pair.",
        "properties": {
          "max_matches": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Max Matches",
            "type": "integer"
          },
          "performance_version_id": {
            "format": "uuid",
            "title": "Performance Version Id",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "source_version_id": {
            "format": "uuid",
            "title": "Source Version Id",
            "type": "string"
          },
          "text": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "source_version_id",
          "performance_version_id",
          "project_id",
          "text"
        ],
        "title": "TextPassageFindWorkflowBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "Workflow": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/WorkflowKind"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "target_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Version Id"
          }
        },
        "required": [
          "project_id",
          "kind"
        ],
        "title": "Workflow",
        "type": "object"
      },
      "WorkflowJobResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "workflow": {
            "$ref": "#/components/schemas/Workflow"
          }
        },
        "required": [
          "workflow",
          "job"
        ],
        "title": "WorkflowJobResponse",
        "type": "object"
      },
      "WorkflowKind": {
        "enum": [
          "understand",
          "correct",
          "compare",
          "create",
          "export"
        ],
        "title": "WorkflowKind",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Persistent projects, immutable music artifacts, asynchronous jobs, and evidence-backed analysis.",
    "title": "Listen Closer music understanding API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workflows/text-passage-find": {
      "post": {
        "description": "Queue CLaMP3 C2 retrieval instead of blocking an API request on model inference.",
        "operationId": "create_text_passage_find_workflow_api_v1_workflows_text_passage_find_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPassageFindWorkflowBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Text Passage Find Workflow"
      }
    },
...
  }
}